Fix hydrate edge cases
This depended on the order of passed pkgs and that we would skip dry nodes already processed.

Fixes #403
