+++ src/types.ts
@@ -29,7 +29,7 @@ export enum Verbosity {
 
 // when we support more variants of these that require specification
 // we will tuple a version in with each eg. 'darwin' | ['windows', 10 | 11 | '*']
-export const SupportedPlatforms = ["darwin", "linux", "windows"] as const
+export const SupportedPlatforms = ["darwin" , "linux" , "windows" , "freebsd" , "netbsd" , "aix" , "solaris" , "illumos"] as const
 export type SupportedPlatform = typeof SupportedPlatforms[number]
 
 export const SupportedArchitectures = ["x86-64", "aarch64"] as const
